+++ src/App.js
@@ -1,71 +1,14 @@
-import { useState } from 'react';
+import Board from './components/Board';
 
-function Square({ value, onSquareClick }) {
+
+export default function App() {
   return (
-    <button className="square" onClick={onSquareClick}>
-      {value}
-    </button>
-  );
-}
-
-export default function Board() {
-  const [currentPlayer, setCurrentPlayer] = useState("X");
-  const [squares, setSquares] = useState(Array(9).fill(null));
-
-  const winner = calculateWinner(squares);
-  const isDraw = squares.every(square => square !== null) && !winner;
-
-  const status = winner
-    ? `Winner: ${winner}`
-    : isDraw
-      ? "Draw"
-      : `Next player: ${currentPlayer}`;
-
-  function handleClick(i) {
-    if (squares[i] || winner){
-      return;
-    } 
-    const nextSquares = squares.slice();
-    nextSquares[i] = currentPlayer;
-    setSquares(nextSquares);
-    setCurrentPlayer(currentPlayer === "X" ? "O" : "X");
-  }
-
-  return (
-    <div className="board">
-      <div className="status">{status}</div>
-
-      {[0, 1, 2].map(row => (
-        <div key={row} className="board-row">
-          {squares.slice(row * 3, row * 3 + 3).map((value, i) => (
-            <Square
-              key={row * 3 + i}
-              value={value}
-              onSquareClick={() => handleClick(row * 3 + i)}
-            />
-          ))}
-        </div>
-      ))}
+    <div className="app">
+      <Board />
     </div>
   );
 }
 
-function calculateWinner(squares) {
-  const lines = [
-    [0, 1, 2],
-    [3, 4, 5],
-    [6, 7, 8],
-    [0, 3, 6],
-    [1, 4, 7],
-    [2, 5, 8],
-    [0, 4, 8],
-    [2, 4, 6]
-  ];
-  for (let i = 0; i < lines.length; i++) {
-    const [a, b, c] = lines[i];
-    if (squares[a] && squares[a] === squares[b] && squares[a] === squares[c]) {
-      return squares[a];
-    }
-  }
-  return null;
-}
+
+
+
